common/requirements.txt
boto3
pinject
pydantic
requests
python-telegram-bot
python-dynamodb-lock-whatnick